Share the document-to-ValueTree restore shape as device_state::toValueTree
The native engine's device factory and the load-time hydration each kept a
private copy of the same Node -> ValueTree conversion. One function next to
the schema now owns it, so every consumer builds the restore tree a device
sees through the same code.
